feat(scanner)!: rename pleno-scan to pleno-pii-scanner; English README
Renames the CLI binary, package name, source directory, and Python
module to make the scope explicit:

- packages/scanner -> packages/pii-scanner
- src/pleno_scan -> src/pleno_pii_scanner
- pyproject name: pleno-scan -> pleno-pii-scanner
- console script: pleno-scan -> pleno-pii-scanner
- workspace member path updated in root pyproject.toml
- Dockerfile builder COPY path updated

Also rewrites both READMEs in English with no parenthetical asides:

- Root README: 65 lines, two surfaces sharing the same NER + Presidio
  pipeline, HTTP usage, CLI usage, single entity table, self-host.
- Scanner README: 98 lines, subcommand list, local vs --base-url offload,
  detected entities with default-profile carve-out, verification labels,
  output formats, .plenoignore syntax, flag table.

35/35 tests green. CLI smoke test confirms PERSON / PHONE_NUMBER / etc
detection unchanged.

BREAKING CHANGE: anyone with 'uv run pleno-scan ...' baked into hooks or
CI must switch to 'uv run pleno-pii-scanner ...'.
