+++ README.md
@@ -1,14 +1,14 @@
 # pleno-anonymize
 
-日本語に強い PII 匿名化基盤。2 つの使い方で同じ NER + Presidio パイプラインを共有する。
+Japanese-first PII anonymization. Two surfaces share the same NER and Presidio pipeline.
 
-- **`pleno-anonymize` server** — `/api/analyze` `/api/redact` と OpenAI/Anthropic/Gemini プロキシを公開する HTTP サービス
-- **`pleno-scan` CLI** — リポジトリやコミット履歴を走査して PII を検出する gitleaks/trufflehog 風スキャナー
+- **`pleno-anonymize` server** — HTTP service exposing `/api/analyze`, `/api/redact`, and OpenAI / Anthropic / Gemini proxies.
+- **`pleno-pii-scanner` CLI** — gitleaks- and trufflehog-style scanner that walks repositories and commit history for PII.
 
-エンドポイント: https://pleno-anonymize.fly.dev (`/docs` で API リファレンス)
-プレイグラウンド: https://plenoai.com/pleno-anonymize/playground
+Endpoint: https://pleno-anonymize.fly.dev with API reference at `/docs`.
+Playground: https://plenoai.com/pleno-anonymize/playground
 
-## サービスを使う (HTTP)
+## Use the service over HTTP
 
 ```bash
 curl -X POST https://pleno-anonymize.fly.dev/api/analyze \
@@ -16,50 +16,50 @@ curl -X POST https://pleno-anonymize.fly.dev/api/analyze \
   -d '{"text":"連絡先 山田太郎 090-1234-5678","language":"ja"}'
 ```
 
-LLM プロキシ経由で送ると、プロバイダに到達する前に PII をマスクし、レスポンスで元の値に復元する。
+Routing through an LLM proxy masks PII before the request reaches the provider, then restores the original values in the response.
 
-| エンドポイント | 上流 |
+| Endpoint | Upstream |
 |---|---|
 | `POST /api/openai/*` | OpenAI Chat Completions / Responses |
 | `POST /api/anthropic/*` | Anthropic Messages |
 | `POST /api/gemini/*` | Google Gemini |
 
-## リポジトリをスキャンする (CLI)
+## Scan repositories with the CLI
 
 ```bash
-uv sync                                   # ja_ner_ja モデル含めてセットアップ
-uv run pleno-scan dir ./my-repo           # ディレクトリ
-uv run pleno-scan git ./my-repo           # working tree + commit 履歴
-uv run pleno-scan github owner/repo       # clone してスキャン
-uv run pleno-scan protect                 # pre-commit (staged hunks のみ)
+uv sync                                   # installs scanner and ja_ner_ja model
+uv run pleno-pii-scanner dir ./my-repo           # directory walk
+uv run pleno-pii-scanner git ./my-repo           # working tree plus commit history
+uv run pleno-pii-scanner github owner/repo       # shallow clone, then scan
+uv run pleno-pii-scanner protect                 # pre-commit guard on staged hunks
 ```
 
-サーバを別マシンで動かしている場合は `--base-url` でオフロードできる:
+To offload analysis to a remote pleno-anonymize, pass `--base-url`:
 
 ```bash
-uv run pleno-scan dir ./my-repo --base-url https://pleno-anonymize.fly.dev
+uv run pleno-pii-scanner dir ./my-repo --base-url https://pleno-anonymize.fly.dev
 ```
 
-詳細は [`packages/scanner/README.md`](packages/scanner/README.md)。
+Full reference at [`packages/pii-scanner/README.md`](packages/pii-scanner/README.md).
 
-## 検出できる PII
+## Detected PII
 
-| 種別 | 検出方法 | エンティティ |
+| Class | Backend | Entities |
 |---|---|---|
-| 自由文 | spaCy NER (`ja_ner_ja`) + Presidio | `PERSON` `ADDRESS` `ORGANIZATION` `DATE_OF_BIRTH` `BANK_ACCOUNT` |
-| 構造化 | 正規表現 + checksum (Luhn / マイナンバー / 法人番号) | `PHONE_NUMBER` `MY_NUMBER` `MY_NUMBER_CORPORATE` `CREDIT_CARD` `PASSPORT` `DRIVER_LICENSE` `HEALTH_INSURANCE` `RESIDENCE_CARD` `POSTAL_CODE` `EMAIL_ADDRESS` `IP_ADDRESS` `URL` |
+| Free text | spaCy NER `ja_ner_ja` and Presidio | `PERSON` `ADDRESS` `ORGANIZATION` `DATE_OF_BIRTH` `BANK_ACCOUNT` |
+| Structured | regex plus checksum for Luhn, My Number, and corporate number | `PHONE_NUMBER` `MY_NUMBER` `MY_NUMBER_CORPORATE` `CREDIT_CARD` `PASSPORT` `DRIVER_LICENSE` `HEALTH_INSURANCE` `RESIDENCE_CARD` `POSTAL_CODE` `EMAIL_ADDRESS` `IP_ADDRESS` `URL` |
 
-サーバ・CLI どちらも同じレジストリを読み込むため、同じ入力に対して同じエンティティ集合を返す。
+Server and CLI load the same recognizer registry, so identical input yields identical entity sets.
 
-## セルフホスト
+## Self-host
 
 ```bash
 docker build -t pleno-anonymize .
 docker run -p 8080:8080 pleno-anonymize
 ```
 
-`fly.toml` 同梱。`flyctl deploy --local-only` でそのまま fly.io に上がる。
+`fly.toml` is included. `flyctl deploy --local-only` ships the same image to fly.io.
 
-## ライセンス
+## License
 
 [AGPL-3.0](LICENSE) · [Privacy Policy](docs/PRIVACY.md) · [DPIA](docs/DPIA.md)
