+++ packages/pii-scanner/src/pleno_pii_scanner/verify.py
@@ -13,7 +13,7 @@ from typing import Iterable
 from pleno_recognizers.types import PiiRecognizer
 from pleno_recognizers.validators import validate
 
-from pleno_scan.models import Finding
+from pleno_pii_scanner.models import Finding
 
 
 def _context_keywords(recognizers: Iterable[PiiRecognizer]) -> dict[str, tuple[str, ...]]:
