+++ pyproject.toml
@@ -1,2 +1,2 @@
 [tool.uv.workspace]
-members = ["server", "packages/training", "packages/recognizers", "packages/scanner"]
+members = ["server", "packages/training", "packages/recognizers", "packages/pii-scanner"]
